first linux adaption in netbeans

+++ LogAndPcapSave/LogData.h
@@ -22,7 +22,7 @@ public:
 	long long timestamp_ms;
 	std::string time;
 
-	DWORD pid;
+	unsigned long pid;
 	std::string msg;
 
 };
